feat: segment each function
