+++ cli/bin/index.js
@@ -612,7 +612,7 @@ var require_command = __commonJS({
     var EventEmitter = require("events").EventEmitter;
     var childProcess = require("child_process");
     var path = require("path");
-    var fs2 = require("fs");
+    var fs3 = require("fs");
     var process9 = require("process");
     var { Argument: Argument2, humanReadableArgName } = require_argument();
     var { CommanderError: CommanderError2 } = require_error();
@@ -1043,11 +1043,11 @@ Expecting one of '${allowedValues.join("', '")}'`);
         const sourceExt = [".js", ".ts", ".tsx", ".mjs", ".cjs"];
         function findFile(baseDir, baseName) {
           const localBin = path.resolve(baseDir, baseName);
-          if (fs2.existsSync(localBin))
+          if (fs3.existsSync(localBin))
             return localBin;
           if (sourceExt.includes(path.extname(baseName)))
             return void 0;
-          const foundExt = sourceExt.find((ext) => fs2.existsSync(`${localBin}${ext}`));
+          const foundExt = sourceExt.find((ext) => fs3.existsSync(`${localBin}${ext}`));
           if (foundExt)
             return `${localBin}${foundExt}`;
           return void 0;
@@ -1059,7 +1059,7 @@ Expecting one of '${allowedValues.join("', '")}'`);
         if (this._scriptPath) {
           let resolvedScriptPath;
           try {
-            resolvedScriptPath = fs2.realpathSync(this._scriptPath);
+            resolvedScriptPath = fs3.realpathSync(this._scriptPath);
           } catch (err) {
             resolvedScriptPath = this._scriptPath;
           }
@@ -7121,7 +7121,300 @@ var {
 } = import_index.default;
 
 // src/index.ts
+var fs2 = __toESM(require("fs"));
+
+// src/utils/installers.ts
 var fs = __toESM(require("fs"));
+var createPackageJson = (libName, includeTailwind) => {
+  fs.writeFileSync(
+    `${libName}/package.json`,
+    JSON.stringify(
+      {
+        name: libName,
+        version: "0.1.0",
+        description: "A simple UI library generated by Skeletor.",
+        main: "dist/cjs/index.js",
+        module: "dist/esm/index.js",
+        files: ["dist"],
+        types: "./dist/index.d.ts",
+        scripts: {
+          build: "rimraf dist && tsc --emitDeclarationOnly && NODE_ENV=production rollup --config",
+          tsc: "tsc"
+        },
+        author: "skeletor-cli",
+        license: "UNLICENSED",
+        devDependencies: __spreadValues({
+          "@babel/core": "^7.14.3",
+          "@babel/preset-env": "^7.14.2",
+          "@babel/preset-react": "^7.13.13",
+          "@rollup/plugin-babel": "^5.3.0",
+          "@rollup/plugin-commonjs": "^17.1.0",
+          "@rollup/plugin-node-resolve": "^11.2.1",
+          "@rollup/plugin-typescript": "^8.2.1",
+          "@types/react": "^17.0.3",
+          "@types/react-dom": "^17.0.3",
+          "babel-loader": "^8.2.2",
+          react: "^18.2.0",
+          "react-dom": "^18.2.0",
+          typescript: "^4.3.5",
+          rimraf: "^3.0.2",
+          rollup: "^2.52.1",
+          "rollup-plugin-peer-deps-external": "^2.2.4",
+          "rollup-plugin-dts": "^3.0.1",
+          webpack: "^5.52.1"
+        }, includeTailwind && {
+          tailwindcss: "^2.2.19",
+          postcss: "^8.3.11",
+          autoprefixer: "^10.3.7",
+          "rollup-plugin-postcss": "^4.0.0"
+        })
+      },
+      null,
+      2
+    )
+  );
+};
+var createTsConfig = (libName) => {
+  fs.writeFileSync(
+    `${libName}/tsconfig.json`,
+    JSON.stringify(
+      {
+        compilerOptions: {
+          target: "es5",
+          jsx: "react",
+          module: "esnext",
+          rootDir: "./src",
+          lib: ["es6", "dom", "es2016", "es2017"],
+          moduleResolution: "node",
+          resolveJsonModule: true,
+          allowJs: false,
+          declaration: true,
+          declarationDir: "types",
+          sourceMap: true,
+          outDir: "./dist",
+          allowSyntheticDefaultImports: true,
+          esModuleInterop: true,
+          forceConsistentCasingInFileNames: true,
+          strict: true,
+          skipLibCheck: true,
+          emitDeclarationOnly: true
+        },
+        include: ["src/**/*"],
+        exclude: ["node_modules", "dist"]
+      },
+      null,
+      2
+    )
+  );
+};
+var createFolderStructure = (libName, includeTailwind) => {
+  fs.mkdirSync(`${libName}/src`);
+  fs.mkdirSync(`${libName}/src/utils`);
+  fs.mkdirSync(`${libName}/src/components`);
+  fs.mkdirSync(`${libName}/src/components/Button`);
+  if (includeTailwind) {
+    fs.mkdirSync(`${libName}/src/styles`);
+  }
+};
+var createIndexFile = (libName, includeTailwind) => {
+  fs.writeFileSync(
+    `${libName}/src/index.ts`,
+    `
+    ${includeTailwind && `import './styles/globals.css';`}
+    export { default as Button, ButtonProps } from './components/Button/index.tsx';
+    `
+  );
+};
+var createButtonComponent = (libName) => {
+  fs.writeFileSync(
+    `${libName}/src/components/Button/index.tsx`,
+    `
+    import React from 'react';
+
+    export interface ButtonProps {
+      label: string;
+    }
+
+    export const Button = ({ label }: ButtonProps) => {
+      return (
+        <button>{label}</button>
+      );
+    };
+  `
+  );
+};
+var createGlobalsCss = (libName) => {
+  fs.writeFileSync(
+    `${libName}/src/styles/globals.css`,
+    `
+    @tailwind base;
+    @tailwind components;
+    @tailwind utilities;
+  `
+  );
+};
+var createClassNamesUtil = (libName) => {
+  fs.writeFileSync(
+    `${libName}/src/utils/classnames.ts`,
+    `
+    export function classnames(...classes: unknown[]) {
+      return classes.filter(Boolean).join(' ');
+    }
+  `
+  );
+};
+var createRollupConfig = (libName, includeTailwind) => {
+  fs.writeFileSync(
+    `${libName}/rollup.config.js`,
+    `
+    import resolve from '@rollup/plugin-node-resolve';
+    import commonjs from '@rollup/plugin-commonjs';
+    import external from 'rollup-plugin-peer-deps-external';
+    import typescript from '@rollup/plugin-typescript';
+    import dts from 'rollup-plugin-dts';
+    import babel from '@rollup/plugin-babel';
+    import pkg from './package.json';
+    ${includeTailwind && `import postcss from "rollup-plugin-postcss";`}
+
+    export default [
+      {
+        input: ['src/index.ts'],
+        output: [
+          {
+            file: pkg.main,
+            format: 'cjs',
+            sourcemap: true,
+            name: 'react-ts-lib',
+          },
+          {
+            file: pkg.module,
+            format: 'esm',
+            sourcemap: true,
+          },
+        ],
+        plugins: [
+          babel({
+            babelHelpers: 'bundled',
+          }),
+          external(),
+          resolve(),
+          commonjs(),
+          ${includeTailwind && `postcss({ 
+            extensions: ['.css'],
+              minimize: true,
+              inject: {
+                insertAt: 'top',
+              },
+          }),`}
+          typescript({
+            tsconfig: './tsconfig.json',
+          }),
+        ],
+        external: ['react', 'react-dom'],
+      },
+      {
+        input: 'dist/esm/types/index.d.ts',
+        output: [
+          {
+            file: 'dist/index.d.ts',
+            format: 'esm',
+          },
+        ],
+        external: [/.css$/],
+        plugins: [dts()],
+      },
+    ];`
+  );
+};
+var createTailwindConfig = (libName) => {
+  fs.writeFileSync(
+    `${libName}/tailwind.config.js`,
+    `
+    module.exports = {
+      purge: ['./src/**/*.{js,jsx,ts,tsx}'],
+      darkMode: false, // or 'media' or 'class'
+      theme: {
+        extend: {},
+      },
+      variants: {
+        extend: {},
+      },
+      plugins: [],
+    }; 
+  `
+  );
+};
+var createPostCssConfig = (libName) => {
+  fs.writeFileSync(
+    `${libName}/postcss.config.js`,
+    `
+    module.exports = {
+      plugins: {
+        tailwindcss: {},
+        autoprefixer: {},
+      },
+    };
+  `
+  );
+};
+var createGitIgnore = (libName) => {
+  fs.writeFileSync(
+    `${libName}/.gitignore`,
+    `# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+
+  # dependencies
+  /node_modules
+  /.pnp
+  .pnp.js
+  
+  # testing
+  /coverage
+  
+  # next.js
+  /.next/
+  /out/
+  
+  # production
+  /build
+  
+  # misc
+  .DS_Store
+  *.pem
+  
+  # debug
+  npm-debug.log*
+  yarn-debug.log*
+  yarn-error.log*
+  
+  # local env files
+  .env.local
+  .env.development.local
+  .env.test.local
+  .env.production.local
+  
+  # vercel
+  .vercel
+`
+  );
+};
+var createBabelConfig = (libName) => {
+  fs.writeFileSync(
+    `${libName}/.babelrc`,
+    `
+    {
+      "presets": ["@babel/preset-env", "@babel/preset-react"]
+    }
+  `
+  );
+};
+var createReadme = (libName) => {
+  fs.writeFileSync(
+    `${libName}/README.md`,
+    `# ${libName}
+
+    A simple UI library generated by Skeletor. \u{1F480}`
+  );
+};
 
 // ../node_modules/ora/index.js
 var import_node_process7 = __toESM(require("node:process"), 1);
@@ -8526,12 +8819,6 @@ program2.description(
 );
 program2.option("--verbose", "Verbose logging.");
 program2.version("0.1.0", "-v, --version", "Logs current version.");
-program2.command("generate:tokens").action(() => __async(exports, null, function* () {
-  updateSpinnerText("\u{1F480} Skeletor is...");
-  yield new Promise((resolve) => setTimeout(resolve, 3e3));
-  spinnerSuccess();
-  console.log("wow your tokens!");
-}));
 program2.command("generate:ui").option("--with-tailwind", "Adds Tailwind CSS to your package.").action((options) => __async(exports, null, function* () {
   updateSpinnerText("\u{1F480} Skeletor is generating your package...");
   yield new Promise((resolve) => setTimeout(resolve, 2e3));
@@ -8539,244 +8826,37 @@ program2.command("generate:ui").option("--with-tailwind", "Adds Tailwind CSS to 
   const includeTailwind = options.withTailwind;
   updateSpinnerText("\u{1F4C1} Setting up your package directory...");
   yield new Promise((resolve) => setTimeout(resolve, 2e3));
-  fs.mkdirSync(libName);
-  updateSpinnerText("\u{1F3D7}  Spinning up a package.json...");
+  fs2.mkdirSync(libName);
+  updateSpinnerText("\u{1F635}\u200D\u{1F4AB} Spinning up a package.json...");
   yield new Promise((resolve) => setTimeout(resolve, 2e3));
-  fs.writeFileSync(
-    `${libName}/package.json`,
-    JSON.stringify(
-      {
-        name: libName,
-        version: "0.1.0",
-        description: "A simple UI library generated by Skeletor.",
-        main: "dist/cjs/index.js",
-        module: "dist/esm/index.js",
-        files: ["dist"],
-        types: "./dist/index.d.ts",
-        scripts: {
-          build: "rimraf dist && tsc --emitDeclarationOnly && NODE_ENV=production rollup --config",
-          tsc: "tsc"
-        },
-        author: "skeletor-cli",
-        license: "UNLICENSED",
-        devDependencies: __spreadValues({
-          "@babel/core": "^7.14.3",
-          "@babel/preset-env": "^7.14.2",
-          "@babel/preset-react": "^7.13.13",
-          "@rollup/plugin-babel": "^5.3.0",
-          "@rollup/plugin-commonjs": "^17.1.0",
-          "@rollup/plugin-node-resolve": "^11.2.1",
-          "@rollup/plugin-typescript": "^8.2.1",
-          "@types/react": "^17.0.3",
-          "@types/react-dom": "^17.0.3",
-          "babel-loader": "^8.2.2",
-          react: "^18.2.0",
-          "react-dom": "^18.2.0",
-          typescript: "^4.3.5",
-          rimraf: "^3.0.2",
-          rollup: "^2.52.1",
-          "rollup-plugin-peer-deps-external": "^2.2.4",
-          "rollup-plugin-dts": "^3.0.1",
-          webpack: "^5.52.1"
-        }, includeTailwind && {
-          tailwindcss: "^2.2.19",
-          postcss: "^8.3.11",
-          autoprefixer: "^10.3.7",
-          "rollup-plugin-postcss": "^4.0.0"
-        })
-      },
-      null,
-      2
-    )
-  );
+  createPackageJson(libName, includeTailwind);
   updateSpinnerText("\u{1F480} Adding a tsconfig too...");
   yield new Promise((resolve) => setTimeout(resolve, 2e3));
-  fs.writeFileSync(
-    `${libName}/tsconfig.json`,
-    JSON.stringify(
-      {
-        compilerOptions: {
-          target: "es5",
-          jsx: "react",
-          module: "esnext",
-          rootDir: "./src",
-          lib: ["es6", "dom", "es2016", "es2017"],
-          moduleResolution: "node",
-          resolveJsonModule: true,
-          allowJs: false,
-          declaration: true,
-          declarationDir: "types",
-          sourceMap: true,
-          outDir: "./dist",
-          allowSyntheticDefaultImports: true,
-          esModuleInterop: true,
-          forceConsistentCasingInFileNames: true,
-          strict: true,
-          skipLibCheck: true,
-          emitDeclarationOnly: true
-        },
-        include: ["src/**/*"],
-        exclude: ["node_modules", "dist"]
-      },
-      null,
-      2
-    )
-  );
-  console.log("Creating directories...");
+  createTsConfig(libName);
   updateSpinnerText("\u{1F477}\u200D\u2640\uFE0F Crafting your first components...");
   yield new Promise((resolve) => setTimeout(resolve, 4e3));
-  fs.mkdirSync(`${libName}/src`);
-  fs.mkdirSync(`${libName}/src/components`);
-  fs.mkdirSync(`${libName}/src/components/Button`);
-  if (includeTailwind) {
-    fs.mkdirSync(`${libName}/src/styles`);
-  }
-  fs.writeFileSync(
-    `${libName}/src/index.ts`,
-    `
-      ${includeTailwind && `import './styles/globals.css';`}
-      export { default as Button, ButtonProps } from './components/Button/index.tsx';
-      `
-  );
-  fs.writeFileSync(
-    `${libName}/src/components/Button/index.tsx`,
-    `
-      import React from 'react';
-
-      export interface ButtonProps {
-        label: string;
-      }
-
-      export const Button = ({ label }: ButtonProps) => {
-        return (
-          <button>{label}</button>
-        );
-      };
-    `
-  );
-  if (includeTailwind) {
-    fs.writeFileSync(
-      `${libName}/src/styles/globals.css`,
-      `
-        @tailwind base;
-        @tailwind components;
-        @tailwind utilities;
-      `
-    );
-  }
+  createFolderStructure(libName, includeTailwind);
+  createIndexFile(libName, includeTailwind);
+  createButtonComponent(libName);
+  createClassNamesUtil(libName);
+  if (includeTailwind)
+    createGlobalsCss(libName);
   updateSpinnerText("\u{1F9FB} Adding a rollup.config.js...");
   yield new Promise((resolve) => setTimeout(resolve, 2e3));
-  fs.writeFileSync(
-    `${libName}/rollup.config.js`,
-    `
-      import resolve from '@rollup/plugin-node-resolve';
-      import commonjs from '@rollup/plugin-commonjs';
-      import external from 'rollup-plugin-peer-deps-external';
-      import typescript from '@rollup/plugin-typescript';
-      import dts from 'rollup-plugin-dts';
-      import babel from '@rollup/plugin-babel';
-      import pkg from './package.json';
-      ${includeTailwind && `import postcss from "rollup-plugin-postcss";`}
-
-      export default [
-        {
-          input: ['src/index.ts'],
-          output: [
-            {
-              file: pkg.main,
-              format: 'cjs',
-              sourcemap: true,
-              name: 'react-ts-lib',
-            },
-            {
-              file: pkg.module,
-              format: 'esm',
-              sourcemap: true,
-            },
-          ],
-          plugins: [
-            babel({
-              babelHelpers: 'bundled',
-            }),
-            external(),
-            resolve(),
-            commonjs(),
-            ${includeTailwind && `postcss({ 
-              extensions: ['.css'],
-                minimize: true,
-                inject: {
-                  insertAt: 'top',
-                },
-            }),`}
-            typescript({
-              tsconfig: './tsconfig.json',
-            }),
-          ],
-          external: ['react', 'react-dom'],
-        },
-        {
-          input: 'dist/esm/types/index.d.ts',
-          output: [
-            {
-              file: 'dist/index.d.ts',
-              format: 'esm',
-            },
-          ],
-          external: [/.css$/],
-          plugins: [dts()],
-        },
-      ];`
-  );
+  createRollupConfig(libName, includeTailwind);
   if (includeTailwind) {
-    updateSpinnerText("\u{1F4A8}  Adding TailwindCSS...");
+    updateSpinnerText("\u{1F4A8} Setting up TailwindCSS...");
     yield new Promise((resolve) => setTimeout(resolve, 4e3));
-    fs.writeFileSync(
-      `${libName}/postcss.config.js`,
-      `
-        module.exports = {
-          plugins: {
-            tailwindcss: {},
-            autoprefixer: {},
-          },
-        };
-      `
-    );
-    fs.writeFileSync(
-      `${libName}/tailwind.config.js`,
-      `
-        module.exports = {
-          purge: ['./src/**/*.{js,jsx,ts,tsx}'],
-          darkMode: false, // or 'media' or 'class'
-          theme: {
-            extend: {},
-          },
-          variants: {
-            extend: {},
-          },
-          plugins: [],
-        }; 
-      `
-    );
+    createTailwindConfig(libName);
+    createPostCssConfig(libName);
   }
   updateSpinnerText("\u{1F480}  Wrapping up!");
   yield new Promise((resolve) => setTimeout(resolve, 2e3));
-  fs.writeFileSync(`${libName}/.gitignore`, "node_modules");
-  fs.writeFileSync(
-    `${libName}/.babelrc`,
-    `
-      {
-        "presets": ["@babel/preset-env", "@babel/preset-react"]
-      }
-    `
-  );
+  createGitIgnore(libName);
+  createBabelConfig(libName);
   updateSpinnerText(`\u{1F4C4} Oh, don't forget to add a README.md!`);
   yield new Promise((resolve) => setTimeout(resolve, 2e3));
-  fs.writeFileSync(
-    `${libName}/README.md`,
-    `# ${libName}
-
-      A simple UI library generated by Skeletor. \u{1F480}`
-  );
+  createReadme(libName);
   updateSpinnerText(`\u{1F480} That's it! ~ cd ${libName} && npm install`);
   yield new Promise((resolve) => setTimeout(resolve, 2e3));
   spinnerSuccess();
