+++ cli/src/index.ts
@@ -1,6 +1,22 @@
 #!/usr/bin/env node
 import { Command } from 'commander';
+import { create } from 'domain';
 import * as fs from 'fs';
+import {
+  createPackageJson,
+  createTsConfig,
+  createFolderStructure,
+  createIndexFile,
+  createButtonComponent,
+  createClassNamesUtil,
+  createGlobalsCss,
+  createRollupConfig,
+  createTailwindConfig,
+  createPostCssConfig,
+  createGitIgnore,
+  createBabelConfig,
+  createReadme,
+} from './utils/installers';
 // import { promiseResolve } from './utils/helpers';
 import { updateSpinnerText, spinnerSuccess } from './utils/spinners';
 
@@ -12,14 +28,14 @@ program.description(
 program.option('--verbose', 'Verbose logging.');
 program.version('0.1.0', '-v, --version', 'Logs current version.');
 
-program.command('generate:tokens').action(async () => {
-  updateSpinnerText('💀 Skeletor is...');
-  // await promiseResolve(3000);
-  await new Promise((resolve) => setTimeout(resolve, 3000));
+// program.command('generate:tokens').action(async () => {
+//   updateSpinnerText('💀 Skeletor is...');
+//   // await promiseResolve(3000);
+//   await new Promise((resolve) => setTimeout(resolve, 3000));
 
-  spinnerSuccess();
-  console.log('wow your tokens!');
-});
+//   spinnerSuccess();
+//   console.log('wow your tokens!');
+// });
 
 program
   .command('generate:ui')
@@ -39,269 +55,55 @@ program
     fs.mkdirSync(libName);
 
     // 2. Create a package.json file
-    updateSpinnerText('🏗  Spinning up a package.json...');
+    updateSpinnerText('😵‍💫 Spinning up a package.json...');
     await new Promise((resolve) => setTimeout(resolve, 2000));
 
-    fs.writeFileSync(
-      `${libName}/package.json`,
-      JSON.stringify(
-        {
-          name: libName,
-          version: '0.1.0',
-          description: 'A simple UI library generated by Skeletor.',
-          main: 'dist/cjs/index.js',
-          module: 'dist/esm/index.js',
-          files: ['dist'],
-          types: './dist/index.d.ts',
-          scripts: {
-            build:
-              'rimraf dist && tsc --emitDeclarationOnly && NODE_ENV=production rollup --config',
-            tsc: 'tsc',
-          },
-          author: 'skeletor-cli',
-          license: 'UNLICENSED',
-          devDependencies: {
-            '@babel/core': '^7.14.3',
-            '@babel/preset-env': '^7.14.2',
-            '@babel/preset-react': '^7.13.13',
-            '@rollup/plugin-babel': '^5.3.0',
-            '@rollup/plugin-commonjs': '^17.1.0',
-            '@rollup/plugin-node-resolve': '^11.2.1',
-            '@rollup/plugin-typescript': '^8.2.1',
-            '@types/react': '^17.0.3',
-            '@types/react-dom': '^17.0.3',
-            'babel-loader': '^8.2.2',
-            react: '^18.2.0',
-            'react-dom': '^18.2.0',
-            typescript: '^4.3.5',
-            rimraf: '^3.0.2',
-            rollup: '^2.52.1',
-            'rollup-plugin-peer-deps-external': '^2.2.4',
-            'rollup-plugin-dts': '^3.0.1',
-            webpack: '^5.52.1',
-            ...(includeTailwind && {
-              tailwindcss: '^2.2.19',
-              postcss: '^8.3.11',
-              autoprefixer: '^10.3.7',
-              'rollup-plugin-postcss': '^4.0.0',
-            }),
-          },
-        },
-        null,
-        2,
-      ),
-    );
+    createPackageJson(libName, includeTailwind);
 
     // 3. Create a tsconfig.json file
     updateSpinnerText('💀 Adding a tsconfig too...');
     await new Promise((resolve) => setTimeout(resolve, 2000));
 
-    fs.writeFileSync(
-      `${libName}/tsconfig.json`,
-      JSON.stringify(
-        {
-          compilerOptions: {
-            target: 'es5',
-            jsx: 'react',
-            module: 'esnext',
-            rootDir: './src',
-            lib: ['es6', 'dom', 'es2016', 'es2017'],
-            moduleResolution: 'node',
-            resolveJsonModule: true,
-            allowJs: false,
-            declaration: true,
-            declarationDir: 'types',
-            sourceMap: true,
-            outDir: './dist',
-            allowSyntheticDefaultImports: true,
-            esModuleInterop: true,
-            forceConsistentCasingInFileNames: true,
-            strict: true,
-            skipLibCheck: true,
-            emitDeclarationOnly: true,
-          },
-          include: ['src/**/*'],
-          exclude: ['node_modules', 'dist'],
-        },
-        null,
-        2,
-      ),
-    );
+    createTsConfig(libName);
 
     // 4. Create directories...
-    console.log('Creating directories...');
     updateSpinnerText('👷‍♀️ Crafting your first components...');
     await new Promise((resolve) => setTimeout(resolve, 4000));
 
-    fs.mkdirSync(`${libName}/src`);
-    fs.mkdirSync(`${libName}/src/components`);
-    fs.mkdirSync(`${libName}/src/components/Button`);
-    if (includeTailwind) {
-      fs.mkdirSync(`${libName}/src/styles`);
-    }
+    createFolderStructure(libName, includeTailwind);
 
     // 5. Create an index.ts file
-    fs.writeFileSync(
-      `${libName}/src/index.ts`,
-      `
-      ${includeTailwind && `import './styles/globals.css';`}
-      export { default as Button, ButtonProps } from './components/Button/index.tsx';
-      `,
-    );
-    fs.writeFileSync(
-      `${libName}/src/components/Button/index.tsx`,
-      `
-      import React from 'react';
-
-      export interface ButtonProps {
-        label: string;
-      }
-
-      export const Button = ({ label }: ButtonProps) => {
-        return (
-          <button>{label}</button>
-        );
-      };
-    `,
-    );
-    if (includeTailwind) {
-      fs.writeFileSync(
-        `${libName}/src/styles/globals.css`,
-        `
-        @tailwind base;
-        @tailwind components;
-        @tailwind utilities;
-      `,
-      );
-    }
+    createIndexFile(libName, includeTailwind);
+    createButtonComponent(libName);
+    createClassNamesUtil(libName);
+    if (includeTailwind) createGlobalsCss(libName);
 
     // 6. Create a rollup.config.js file
     updateSpinnerText('🧻 Adding a rollup.config.js...');
     await new Promise((resolve) => setTimeout(resolve, 2000));
 
-    fs.writeFileSync(
-      `${libName}/rollup.config.js`,
-      `
-      import resolve from '@rollup/plugin-node-resolve';
-      import commonjs from '@rollup/plugin-commonjs';
-      import external from 'rollup-plugin-peer-deps-external';
-      import typescript from '@rollup/plugin-typescript';
-      import dts from 'rollup-plugin-dts';
-      import babel from '@rollup/plugin-babel';
-      import pkg from './package.json';
-      ${includeTailwind && `import postcss from "rollup-plugin-postcss";`}
-
-      export default [
-        {
-          input: ['src/index.ts'],
-          output: [
-            {
-              file: pkg.main,
-              format: 'cjs',
-              sourcemap: true,
-              name: 'react-ts-lib',
-            },
-            {
-              file: pkg.module,
-              format: 'esm',
-              sourcemap: true,
-            },
-          ],
-          plugins: [
-            babel({
-              babelHelpers: 'bundled',
-            }),
-            external(),
-            resolve(),
-            commonjs(),
-            ${
-              includeTailwind &&
-              `postcss({ 
-              extensions: ['.css'],
-                minimize: true,
-                inject: {
-                  insertAt: 'top',
-                },
-            }),`
-            }
-            typescript({
-              tsconfig: './tsconfig.json',
-            }),
-          ],
-          external: ['react', 'react-dom'],
-        },
-        {
-          input: 'dist/esm/types/index.d.ts',
-          output: [
-            {
-              file: 'dist/index.d.ts',
-              format: 'esm',
-            },
-          ],
-          external: [/\.css$/],
-          plugins: [dts()],
-        },
-      ];`,
-    );
+    createRollupConfig(libName, includeTailwind);
 
     // 7. Create extra files
     if (includeTailwind) {
-      updateSpinnerText('💨  Adding TailwindCSS...');
+      updateSpinnerText('💨 Setting up TailwindCSS...');
       await new Promise((resolve) => setTimeout(resolve, 4000));
 
-      fs.writeFileSync(
-        `${libName}/postcss.config.js`,
-        `
-        module.exports = {
-          plugins: {
-            tailwindcss: {},
-            autoprefixer: {},
-          },
-        };
-      `,
-      );
-
-      fs.writeFileSync(
-        `${libName}/tailwind.config.js`,
-        `
-        module.exports = {
-          purge: ['./src/**/*.{js,jsx,ts,tsx}'],
-          darkMode: false, // or 'media' or 'class'
-          theme: {
-            extend: {},
-          },
-          variants: {
-            extend: {},
-          },
-          plugins: [],
-        }; 
-      `,
-      );
+      createTailwindConfig(libName);
+      createPostCssConfig(libName);
     }
 
     updateSpinnerText('💀  Wrapping up!');
     await new Promise((resolve) => setTimeout(resolve, 2000));
 
-    fs.writeFileSync(`${libName}/.gitignore`, 'node_modules');
-    fs.writeFileSync(
-      `${libName}/.babelrc`,
-      `
-      {
-        "presets": ["@babel/preset-env", "@babel/preset-react"]
-      }
-    `,
-    );
+    createGitIgnore(libName);
+    createBabelConfig(libName);
 
     // 8. Create a README.md file
     updateSpinnerText(`📄 Oh, don't forget to add a README.md!`);
     await new Promise((resolve) => setTimeout(resolve, 2000));
 
-    fs.writeFileSync(
-      `${libName}/README.md`,
-      `# ${libName}
-
-      A simple UI library generated by Skeletor. 💀`,
-    );
+    createReadme(libName);
 
     updateSpinnerText(`💀 That's it! ~ cd ${libName} && npm install`);
     await new Promise((resolve) => setTimeout(resolve, 2000));
